3 add changes to work also with v1725

+++ fadc/src/include/TimeFunctions.h
@@ -1,7 +1,7 @@
 #ifndef TIME_FUNCTION_HGUARD
 #define TIME_FUNCTION_HGUARD
 
-double GetTimeFADC250Firmware(int *fadc,double *par);
+double GetTimeFADC250Firmware(unsigned short *fadc,double *par);
 
 
 #endif
